backend/requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
python-dotenv==1.0.0
langchain==0.1.11
langchain-groq==0.0.1
chromadb==0.4.22
pandas==2.2.0
python-multipart==0.0.9
pydantic==2.6.1
requests==2.31.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.27
alembic==1.13.1
openfoodfacts==0.1.3 